# services/api-gateway/src/main/resources/static/openapi/flowpay-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FlowPay API",
    "version": "v1",
    "description": "FlowPay is a digital wallet platform. This is the **public API**, reached through the gateway at\n`http://localhost:8080`.\n\n## Quick start\n1. **Register** - `POST /api/v1/users`\n2. **Log in** - `POST /api/v1/auth/login` -> copy the `accessToken`\n3. Click **Authorize** (top right) and paste the token. Protected endpoints (padlock icon) now work.\n4. **View / update your profile** - `GET` / `PATCH /api/v1/users/{id}`\n\n## How security works\n* Only registration, login, refresh, logout, password reset and the health check are public.\n* Everything else needs `Authorization: Bearer <accessToken>`. Tokens are JWTs signed with RS256 and last 15 minutes;\n  renew them with `POST /api/v1/auth/refresh` (refresh tokens are single use).\n* You only ever see and change **your own** data. Another customer's id returns `404`.\n* Several endpoints are rate limited per IP address and return `429` with a `Retry-After` header when exceeded.\n* Every error has the same shape: `{timestamp, status, error, message, path}`. Branch on `error`, not `message`.\n\n## Emails in development\nPassword reset and two-step verification send real emails to a local inbox (Mailpit): http://localhost:8025.\n"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local gateway (docker compose)"
    }
  ],
...
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "2. Authentication"
        ],
        "summary": "Log in with email and password",
        "description": "Signs a customer in. **Public.** Limited to 10 attempts per minute per IP address (shared with `login/verify`).\n\n### How it works\n* Correct password, two-step verification **off** -> `200` with `accessToken` (JWT, 15 min) and `refreshToken` (7 days).\n* Correct password, two-step verification **on** -> `200` with `{\"mfaRequired\": true, \"challengeId\": ...}` and **no tokens**. A 6-digit code is emailed; finish with `POST /api/v1/auth/login/verify`.\n* Anything else -> the **same** `401 INVALID_CREDENTIALS` message for a wrong password, an unknown email and a locked account, so the API never reveals whether an account exists.\n* **5 wrong passwords lock the account for 15 minutes** (even the correct password is refused while locked).\n\n### How to test\n1. Register a customer first (`POST /api/v1/users`).\n2. Log in with the right password. Expect **200** and tokens. Copy `accessToken` into the Authorize box.\n3. Log in with a wrong password. Expect **401 INVALID_CREDENTIALS**. Try an unknown email: the response is identical.\n4. Fail 5 times in a row, then use the correct password. Still **401** (locked for 15 minutes; a password reset unlocks it).\n5. Hit it 11 times in a minute. Expect **429**.\n6. To try the two-step response, enable it under \"4. Two-step verification\" and log in again.\n",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Tokens, or an MFA challenge when two-step verification is enabled.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/TokenResponse"
                    },
                    {
                      "$ref": "#/components/schemas/MfaChallengeResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "INVALID_CREDENTIALS (wrong password, unknown email, or locked account).",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMIT_EXCEEDED.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "MFA_UNAVAILABLE: two-step is on but the code email could not be sent. No tokens issued.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login/verify": {
      "post": {
        "tags": [
          "2. Authentication"
        ],
        "summary": "Finish login with the emailed code",
        "description": "Second step of login for customers who turned on two-step verification. **Public.** Shares the login rate limit.\n\n### How it works\nThe code is 6 digits, valid **5 minutes**, usable **once**, with **5 guesses**. Wrong guesses also count towards the account lockout.\nEvery failure (wrong code, expired, already used, unknown challenge) returns the same `401 INVALID_MFA_CODE`.\n\n### How to test\n1. Enable two-step verification, then call `POST /api/v1/auth/login`. Copy `challengeId`.\n2. Open the inbox (Mailpit at http://localhost:8025) and read the 6-digit code.\n3. Send `challengeId` + `code`. Expect **200** with tokens.\n4. Send the same code again. Expect **401 INVALID_MFA_CODE** (single use).\n5. Request a new login code and send `000000` five times. The code is now dead, even the right one.\n6. A malformed code such as `12ab` returns **400**.\n",
        "operationId": "verifyLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login complete.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR: challengeId or code malformed.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "INVALID_MFA_CODE.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMIT_EXCEEDED.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "2. Authentication"
        ],
        "summary": "Log out this session",
        "description": "Ends the session that the refresh token belongs to. **Public** (the refresh token is the proof).\n\n### How it works\nThe refresh token (and every token in its chain) is revoked. Calling it again, or with an unknown token, still returns `204`, so it is safe to repeat and reveals nothing.\nThe access token already issued stays valid until it expires (at most 15 minutes).\n\n### How to test\n1. Log in and copy `refreshToken`.\n2. Call logout with it. Expect **204**.\n3. Call refresh with the same token. Expect **401**.\n4. Call logout again. Still **204**.\n",
        "operationId": "logout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Session ended (or there was nothing to end)."
          },
          "400": {
            "description": "VALIDATION_ERROR.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout-all": {
      "post": {
        "tags": [
          "2. Authentication"
        ],
        "summary": "Log out everywhere",
        "description": "Ends **every** session of the signed-in customer (phone, laptop, ...). **Requires a bearer token.**\n\n### How it works\nAll of the customer's refresh tokens are revoked. Existing access tokens expire on their own within 15 minutes.\n\n### How to test\n1. Log in twice to create two sessions; keep both `refreshToken` values and one `accessToken`.\n2. Call this endpoint with the access token. Expect **204**.\n3. Call refresh with each refresh token. Both return **401**.\n4. Call it without a token. Expect **401 UNAUTHENTICATED**.\n",
        "operationId": "logoutAll",
        "responses": {
          "204": {
            "description": "All sessions ended."
          },
          "401": {
            "description": "UNAUTHENTICATED.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "2. Authentication"
        ],
        "summary": "Get a new access token",
        "description": "Exchanges a refresh token for a fresh access token **and a new refresh token**. **Public** (the refresh token is the proof).\n\n### How it works\nRefresh tokens are **single use** (\"rotation\"). Using one returns a new pair and retires the old token.\nIf an already-used token is presented again it may have been stolen, so the **whole session is revoked**: every token in that chain stops working.\n\n### How to test\n1. Log in and copy `refreshToken`.\n2. Call refresh with it. Expect **200** and a **different** `refreshToken`.\n3. Call refresh again with the **old** token. Expect **401 INVALID_REFRESH_TOKEN**.\n4. Now try the **new** token from step 2. Also **401**: reuse ended the session. Log in again.\n",
        "operationId": "refreshToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "A new token pair.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "INVALID_REFRESH_TOKEN: unknown, expired, already used, or revoked.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users": {
      "post": {
        "tags": [
          "1. Customers"
        ],
        "summary": "Register a new customer",
        "description": "Creates a FlowPay account. **Public** (no token needed). Limited to 5 requests per minute per IP address.\n\n### How it works\n1. The request is validated (see the field rules in the schema).\n2. The email is lowercased. Email and phone number must each be unique.\n3. The customer is saved with a generated `id` and status `PENDING_VERIFICATION`.\n4. The password is sent to **auth-service**, which hashes it (BCrypt) and stores it. This service never stores or returns it.\n5. If auth-service is unavailable the whole registration is rolled back and you get `503`: there is never an account without a way to log in.\n\n### How to test\n1. Send the example body. Expect **201**, and copy the `id` from the response.\n2. Send the exact same body again. Expect **409 DUPLICATE_CUSTOMER** (same email, in any letter case).\n3. Change only the email and keep the phone number. Still **409** (phone must be unique too).\n4. Send `{}` or a weak password such as `weak`. Expect **400 VALIDATION_ERROR** listing each bad field.\n5. Send 6+ requests within a minute from one machine. Expect **429** with a `Retry-After` header.\n6. Next step: log in with `POST /api/v1/auth/login`.\n",
        "operationId": "registerCustomer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterCustomerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Account created. The Location header points at the profile URL.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponse"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR or MALFORMED_REQUEST.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "DUPLICATE_CUSTOMER: email or phone number already registered.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMIT_EXCEEDED. Wait for Retry-After seconds.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "REGISTRATION_UNAVAILABLE: auth-service could not store the login. Nothing was created; retry.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error body used by every endpoint.",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-30T16:13:30.942Z"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code.",
            "example": 400
          },
          "error": {
            "type": "string",
            "description": "Stable machine-readable code to branch on.",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "description": "Human-readable explanation.",
            "example": "email must be a well-formed email address"
          },
          "path": {
            "type": "string",
            "description": "The request path.",
            "example": "/api/v1/users"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "description": "Credentials used to sign in.",
        "properties": {
          "email": {
            "type": "string",
            "description": "The email used at registration (case-insensitive).",
            "example": "ada.obi@example.com",
            "maxLength": 254,
            "minLength": 0
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "The account password.",
            "example": "Sup3rSecret",
            "maxLength": 128,
            "minLength": 0
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "MfaChallengeResponse": {
        "type": "object",
        "description": "A one-time code was emailed. Submit it with this challengeId to continue.",
        "properties": {
          "mfaRequired": {
            "type": "boolean",
            "description": "Always true: the client must complete a second step.",
            "example": true
          },
          "challengeId": {
            "type": "string",
            "format": "uuid",
            "description": "Identifies the emailed code. Send it with the code.",
            "example": "9636018a-2f90-46a7-add3-2549b9841b5f"
          },
          "expiresIn": {
            "type": "integer",
            "format": "int64",
            "description": "Seconds until the code expires.",
            "example": 300
          }
        }
      },
...
      "RefreshTokenRequest": {
        "type": "object",
        "description": "A refresh token previously issued by login, login/verify or refresh.",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "The opaque refresh token. Each one works only once.",
            "example": "SsCed_IPmKTsFYekt7bVScwtaMIB1whQezv9bXeKxrI",
            "maxLength": 200,
            "minLength": 0
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "RegisterCustomerRequest": {
        "type": "object",
        "description": "Details needed to open a FlowPay account.",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "Given name.",
            "example": "Ada",
            "maxLength": 100,
            "minLength": 0
          },
          "lastName": {
            "type": "string",
            "description": "Family name.",
            "example": "Obi",
            "maxLength": 100,
            "minLength": 0
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address. Stored in lowercase and must be unique. Used to log in.",
            "example": "ada.obi@example.com",
            "maxLength": 254,
            "minLength": 0
          },
          "phoneNumber": {
            "type": "string",
            "description": "Phone number in international format (E.164). Must be unique.",
            "example": "+2348012345678",
            "minLength": 1,
            "pattern": "^\\+[1-9]\\d{7,14}$"
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "8-72 characters with at least one uppercase letter, one lowercase letter and one digit. Hashed by auth-service; never stored or returned by user-service.",
            "example": "Sup3rSecret",
            "maxLength": 72,
            "minLength": 8,
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).*$"
          },
          "country": {
            "type": "string",
            "description": "Country as an ISO 3166-1 alpha-2 code.",
            "example": "NG",
            "minLength": 1,
            "pattern": "^[A-Z]{2}$"
          }
        },
        "required": [
          "country",
          "email",
          "firstName",
          "lastName",
          "password",
          "phoneNumber"
        ]
      },
      "TokenResponse": {
        "type": "object",
        "description": "A signed-in session: a short-lived access token and a single-use refresh token.",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT (RS256). Send as 'Authorization: Bearer <token>'. Valid 15 minutes. The 'sub' claim is the customer id.",
            "example": "eyJraWQiOiJmbG93cGF5LTEiLCJhbGciOiJSUzI1NiJ9..."
          },
          "tokenType": {
            "type": "string",
            "example": "Bearer"
          },
          "expiresIn": {
            "type": "integer",
            "format": "int64",
            "description": "Access token lifetime in seconds.",
            "example": 900
          },
          "refreshToken": {
            "type": "string",
            "description": "Opaque token valid 7 days. Exchange it at /refresh; each one works once.",
            "example": "SsCed_IPmKTsFYekt7bVScwtaMIB1whQezv9bXeKxrI"
          }
        }
      },
...
      "VerifyOtpRequest": {
        "type": "object",
        "description": "The emailed one-time code together with the challenge it belongs to.",
        "properties": {
          "challengeId": {
            "type": "string",
            "format": "uuid",
            "description": "The challengeId returned when the code was requested.",
            "example": "9636018a-2f90-46a7-add3-2549b9841b5f"
          },
          "code": {
            "type": "string",
            "description": "The 6-digit code from the email. Valid 5 minutes, single use, 5 attempts.",
            "example": "904826",
            "minLength": 1,
            "pattern": "^\\d{6}$"
          }
        },
        "required": [
          "challengeId",
          "code"
        ]
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token from POST /api/v1/auth/login (or /login/verify)."
      }
    }
  }
}